Homepage Changes
Lesson 3 - Updating Homepage. In this lesson, we learned to add images from relative folders, the importance of using alt and title tags for search engines, humans, and browser visibility, and classes/ids. Using CSS properties to change the style of the background, text, alignment. Adding a "contact" link that sends an email upon click.